Exception 404 handling, from MR.

+++ src/main/java/ua/com/foxminded/volodymyrtolpiekin/school/service/CourseServiceImpl.java
@@ -1,13 +1,16 @@
 package ua.com.foxminded.volodymyrtolpiekin.school.service;
 
+import lombok.extern.log4j.Log4j2;
 import org.springframework.stereotype.Service;
 import ua.com.foxminded.volodymyrtolpiekin.school.dao.JpaCourseDao;
 import ua.com.foxminded.volodymyrtolpiekin.school.models.Course;
 
+import javax.persistence.EntityNotFoundException;
 import java.util.List;
 import java.util.Optional;
 
 @Service
+@Log4j2
 public class CourseServiceImpl implements CourseService{
 
     private final JpaCourseDao jpaCourseDao;
@@ -18,7 +21,11 @@ public class CourseServiceImpl implements CourseService{
 
     @Override
     public Optional<Course> findById(int id) {
-        return jpaCourseDao.findById(id);
+        Optional<Course> returnedCourse = jpaCourseDao.findById(id);
+        if(returnedCourse.isPresent()) {
+            return returnedCourse;
+        }
+        throw new EntityNotFoundException(String.format("404.Course with id %d not found", id));
     }
 
     @Override
@@ -39,6 +46,11 @@ public class CourseServiceImpl implements CourseService{
 
     @Override
     public void deleteById(int id){
-        jpaCourseDao.deleteById(id);
+        Optional<Course> returnedCourse = jpaCourseDao.findById(id);
+        if(returnedCourse.isPresent()) {
+            jpaCourseDao.deleteById(id);
+        }
+        throw new EntityNotFoundException(String.format("404. Course with id %d not found", id));
+
     }
 }
